Add plugin implementation
